# apps/auth-service/src/main/java/com/auth/dto/LoginRequest.java
package com.auth.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class LoginRequest {
    
    @NotBlank(message = "El email es requerido")
    @Email(message = "El email debe tener un formato válido")
    private String username;
    
    @NotBlank(message = "La contraseña es requerida")
    private String password;
}